moves DateUtil.java
